Rename cloud backup runtime actors
